requirements.txt
aiosqlite
alembic
authlib
databases
fastapi
httpx
ormar
pydantic
python-dotenv
requests
SQLAlchemy
starlette
uvicorn